# pnmc-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PNMC Migration API
  version: 1.0.0
  summary: Contrato API inicial para migracion incremental PNMC
  description: |
    Contrato base para desacoplar integraciones sensibles del frontend actual,
    habilitando una migracion paralela sin tocar la UI existente.
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://api-dev.pnmc.local
    description: Dev (referencia)
...
paths:
...
  /api/v1/news/articles:
    get:
      tags: [noticias]
      summary: Listar articulos de noticias
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - in: query
          name: month
          schema:
            type: string
            pattern: '^[0-9]{4}-(0[1-9]|1[0-2])$'
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: Noticias retornadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedNewsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/news/articles/{articleId}:
    get:
      tags: [noticias]
      summary: Obtener articulo por id
      parameters:
        - $ref: '#/components/parameters/ArticleId'
      responses:
        '200':
          description: Articulo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsArticle'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/participation/submissions:
    get:
      tags: [participacion]
      summary: Listar envios de participacion (vista tabular)
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - in: query
          name: actorType
          schema:
            type: string
            enum: [individual, collective, organization, festival, market, space]
        - in: query
          name: department
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: Listado paginado de envios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedParticipationSubmissionSummaryResponse'
    post:
      tags: [participacion]
      summary: Registrar participacion
      description: |
        Endpoint server-side para reemplazar el guardado desde middleware local de Vite.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipationSubmissionRequest'
      responses:
        '201':
          description: Registro creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipationSubmissionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          description: Error de validacion de dominio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/participation/submissions/{reference}:
    get:
      tags: [participacion]
      summary: Consultar envio por referencia
      parameters:
        - in: path
          name: reference
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Registro encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipationSubmissionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/map-participation:
    post:
      tags: [participacion]
      summary: Endpoint de compatibilidad con flujo legacy
      description: |
        Mantiene el contrato de `POST /api/map-participation` del frontend actual
        para conectar gradualmente sin romper la UI existente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipationSubmissionRequest'
      responses:
        '200':
          description: Registro aceptado en backend de migracion
          content:
            application/json:
              schema:
                type: object
                required: [fileName, message, reference, status, externalSyncStatus]
                properties:
                  fileName: { type: string }
                  message: { type: string }
                  reference: { type: string }
                  status: { type: string }
                  externalSyncStatus: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/admin/data/news/articles:
    post:
      tags: [admin-data]
      summary: Crear o actualizar articulo de noticias en base backend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsArticleUpsertRequest'
      responses:
        '200':
          description: Articulo persistido
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  parameters:
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
        default: 0
...
    ArticleId:
      in: path
      name: articleId
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Solicitud invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    InternalError:
      description: Error interno
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    ProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    Pagination:
      type: object
      required: [limit, offset, total]
      properties:
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer }
...
    NewsArticle:
      type: object
      required: [id, title, date]
      properties:
        id: { type: string }
        date: { type: string }
        category: { type: string }
        title: { type: string }
        summary: { type: string }
        contentHtml: { type: string }
        imageUrl: { type: string }
        source:
          type: string
          enum: [backend_sql]
    PagedNewsResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NewsArticle'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    ParticipationSubmissionRequest:
      type: object
      required:
        - actorType
        - actorName
        - email
        - phone
        - department
        - municipality
        - musicalFields
        - description
        - contribution
        - consent
      properties:
        reference:
          type: string
          description: Si no se envia, el backend la genera.
        actorType:
          type: string
          enum: [individual, collective, organization, festival, market, space]
        actorTypeLabel: { type: string }
        actorName:
          type: string
          minLength: 2
          maxLength: 200
        individualFirstName: { type: string }
        individualLastName: { type: string }
        identificationType: { type: string }
        identificationNumber: { type: string }
        hasArtisticName: { type: boolean }
        artisticName: { type: string }
        responsibleEntity: { type: string }
        contactName: { type: string }
        contactRole: { type: string }
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 7
          maxLength: 32
        department:
          type: string
          minLength: 2
        municipality:
          type: string
          minLength: 2
        territoryScope: { type: string }
        website: { type: string }
        facebookUrl: { type: string }
        instagramUrl: { type: string }
        roles:
          type: array
          items: { type: string }
        musicalFields: { type: string }
        description: { type: string }
        contribution: { type: string }
        needs: { type: string }
        organizationSubtype: { type: string }
        yearFounded: { type: string }
        legalStatus: { type: string }
        mainPrograms: { type: string }
        festivalDurationDays: { type: string }
        festivalSetting: { type: string }
        festivalVenueMode: { type: string }
        festivalAdditionalLocations:
          type: array
          items:
            type: object
            properties:
              department: { type: string }
              municipality: { type: string }
        festivalFrequency: { type: string }
        festivalVersions: { type: string }
        festivalHabitualMonths:
          type: array
          items: { type: string }
        festivalTicketing: { type: string }
        openCall: { type: string }
        festivalThisYearStatus: { type: string }
        festivalThisYearDate: { type: string, format: date }
        festivalThisYearStartDate: { type: string, format: date }
        festivalThisYearEndDate: { type: string, format: date }
        festivalCurrentOpenCall: { type: string }
        festivalOpenCallDeadline: { type: string, format: date }
        marketFrequency: { type: string }
        marketEditionsCount: { type: string }
        averageBuyers: { type: string }
        linkedFestival: { type: string }
        linkedFestivalName: { type: string }
        marketHabitualMonths:
          type: array
          items: { type: string }
        marketThisYearStatus: { type: string }
        marketThisYearMonth: { type: string }
        marketThisYearDate: { type: string, format: date }
        individualProfile: { type: string }
        trajectoryYears: { type: string }
        linkedProcesses: { type: string }
        members: { type: string }
        musicalPractice: { type: string }
        circulationScope: { type: string }
        collectiveTrajectory: { type: string }
        spaceType: { type: string }
        spaceCapacity: { type: string }
        spaceUses: { type: string }
        technicalEquipment: { type: string }
        consent:
          type: boolean
    ParticipationSubmissionResponse:
      type: object
      required: [reference, status, submittedAt, externalSyncStatus, externalSyncMessage]
      properties:
        reference: { type: string }
        status:
          type: string
          enum: [accepted, accepted_with_warnings, rejected]
        submittedAt:
          type: string
          format: date-time
        message: { type: string }
        externalSyncStatus:
          type: string
          enum: [backend_only, not_configured, synced, pending_retry]
        externalSyncMessage: { type: string }
    ParticipationSubmissionSummary:
      type: object
      required: [reference, submittedAt, actorType, actorName, email, department, municipality, externalSyncStatus]
      properties:
        reference: { type: string }
        submittedAt:
          type: string
          format: date-time
        actorType: { type: string }
        actorName: { type: string }
        email: { type: string, format: email }
        department: { type: string }
        municipality: { type: string }
        externalSyncStatus:
          type: string
          enum: [backend_only, not_configured, synced, pending_retry]
...
    NewsArticleUpsertRequest:
      type: object
      required: [title]
      properties:
        id: { type: string }
        date: { type: string }
        category: { type: string }
        title: { type: string }
        summary: { type: string }
        contentHtml: { type: string }
        imageUrl: { type: string }
...
    PagedParticipationSubmissionSummaryResponse:
      type: object
      required: [items, limit, offset, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ParticipationSubmissionSummary'
        limit: { type: integer, minimum: 1 }
        offset: { type: integer, minimum: 0 }
        total: { type: integer, minimum: 0 }